backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
pydantic==2.9.2
pydantic-settings==2.5.2

pymongo==4.10.1

numpy==2.1.3
faiss-cpu==1.9.0.post1

openai==1.57.0

python-dotenv==1.0.1
tenacity==9.0.0
orjson==3.10.12
